+++ src/index.ts
@@ -1,6 +1,6 @@
 import { GraphQLServer } from 'graphql-yoga'
-import { schema } from './schema'
 import { createContext } from './context'
+import { schema } from './schema'
 
 new GraphQLServer({ schema, context: createContext }).start(() =>
   console.log(
