+++ src/schema/resolvers/index.ts
@@ -1,0 +1,5 @@
+import { merge } from 'lodash'
+import Post from './Post'
+import User from './User'
+
+export default merge({}, Post, User)
